[Interpreter] Remove interp path (#849)
* helper fn

* replacement interpretation func

* continued reorg

* flatten folders

* fix bug

* remove cidr specific tests

* clean up comment

* move possible port calculation to only occur once rather than on ceach call to the convergence loop

* seq interpreter run def
