+++ interp/src/interpreter/group_interpreter.rs
@@ -1,8 +1,8 @@
-use super::super::interpret_group::{eval_prims, finish_interpretation};
-use super::super::utils::{self, ConstCell, ConstPort};
+use super::utils::{self, ConstCell, ConstPort};
 use crate::environment::InterpreterState;
 use crate::errors::{InterpreterError, InterpreterResult};
-use crate::utils::{AsRaw, PortAssignment};
+use crate::interpreter::utils::get_dest_cells;
+use crate::utils::{AsRaw, PortAssignment, RcOrConst};
 use crate::values::Value;
 use calyx::ir::{self, Assignment, Cell, RRC};
 use std::cell::Ref;
@@ -100,6 +100,7 @@ pub struct AssignmentInterpreter {
     cont_assigns: iir::ContinuousAssignments,
     cells: Vec<RRC<Cell>>,
     val_changed: Option<bool>,
+    possible_ports: HashSet<*const ir::Port>,
 }
 
 impl AssignmentInterpreter {
@@ -117,6 +118,12 @@ impl AssignmentInterpreter {
             assigns.get_ref().iter().chain(cont_assigns.iter()),
             done_signal,
         );
+        let possible_ports: HashSet<*const ir::Port> = assigns
+            .get_ref()
+            .iter()
+            .chain(cont_assigns.iter())
+            .map(|a| a.dst.as_raw())
+            .collect();
 
         Self {
             state,
@@ -125,6 +132,7 @@ impl AssignmentInterpreter {
             cont_assigns: Rc::clone(cont_assigns),
             cells,
             val_changed: None,
+            possible_ports,
         }
     }
 
@@ -179,14 +187,6 @@ impl AssignmentInterpreter {
     /// converge
     pub fn step_convergence(&mut self) -> InterpreterResult<()> {
         self.val_changed = Some(true); // always run convergence if called
-
-        let possible_ports: HashSet<*const ir::Port> = self
-            .assigns
-            .get_ref()
-            .iter()
-            .chain(self.cont_assigns.iter())
-            .map(|a| a.dst.as_raw())
-            .collect();
 
         // this unwrap is safe
         while self.val_changed.unwrap() {
@@ -245,7 +245,7 @@ impl AssignmentInterpreter {
 
             //now assign rest to 0
             //first get all that need to be 0
-            for port in &possible_ports - &assigned_const_ports {
+            for port in &self.possible_ports - &assigned_const_ports {
                 //need to set to zero, because unassigned
                 //ok now proceed
 
@@ -281,15 +281,6 @@ impl AssignmentInterpreter {
     pub fn step(&mut self) -> InterpreterResult<()> {
         self.step_cycle()?;
         self.step_convergence()
-    }
-
-    /// Run interpreter until it is finished executing and return the output
-    /// environment
-    pub fn run_and_deconstruct(
-        mut self,
-    ) -> InterpreterResult<InterpreterState> {
-        self.run()?;
-        self.deconstruct()
     }
 
     /// Run the interpreter until it finishes executing
@@ -359,4 +350,115 @@ impl AssignmentInterpreter {
     pub fn get_mut_env(&mut self) -> &mut InterpreterState {
         &mut self.state
     }
+}
+
+/// Evaluates the primitives corresponding to the given iterator of cells, based
+/// on the current environment. Returns a set of assignments that may change
+/// based on the updates to primitive values.
+///
+/// Note: this function could be written with only one lifetime, but it is worth
+/// noting that the returned assignments refs are tied to the dependency map and
+/// thus to the assignments it is referencing meanwhile the lifetime on the
+/// given cell RRCs is unrelated and largely irrelevant as the prim_map is keyed
+/// off of port raw pointers whose lifetime is uncoupled from the cells.
+pub(crate) fn eval_prims<'a, 'b, I: Iterator<Item = &'b RRC<ir::Cell>>>(
+    env: &mut InterpreterState,
+    exec_list: I,
+    reset_flag: bool, // reset vals or execute normally
+) -> InterpreterResult<bool> {
+    let mut val_changed = false;
+    // split mutability
+    // TODO: change approach based on new env, once ready
+    let ref_clone = env.cell_map.clone(); // RC clone
+    let mut prim_map = ref_clone.borrow_mut();
+
+    let mut update_list: Vec<(RRC<ir::Port>, Value)> = vec![];
+
+    for cell in exec_list {
+        let inputs = get_inputs(env, &cell.borrow());
+
+        let executable = prim_map.get_mut(&cell.as_raw());
+
+        if let Some(prim) = executable {
+            let new_vals = if reset_flag {
+                prim.reset(&inputs)
+            } else {
+                prim.execute(&inputs)
+            };
+
+            for (port, val) in new_vals? {
+                let port_ref = cell.borrow().find(port).unwrap();
+
+                let current_val = env.get_from_port(&port_ref.borrow());
+
+                if *current_val != val {
+                    val_changed = true;
+                    // defer value update until after all executions
+                    update_list.push((Rc::clone(&port_ref), val));
+                }
+            }
+        }
+    }
+
+    for (port, val) in update_list {
+        env.insert(port, val);
+    }
+
+    Ok(val_changed)
+}
+
+fn get_inputs<'a>(
+    env: &'a InterpreterState,
+    cell: &ir::Cell,
+) -> Vec<(ir::Id, &'a Value)> {
+    cell.ports
+        .iter()
+        .filter_map(|p| {
+            let p_ref: &ir::Port = &p.borrow();
+            match &p_ref.direction {
+                ir::Direction::Input => {
+                    Some((p_ref.name.clone(), env.get_from_port(p_ref)))
+                }
+                _ => None,
+            }
+        })
+        .collect()
+}
+
+/// Concludes interpretation to a group, effectively setting the go signal low
+/// for a given group. This function updates the values in the environment
+/// accordingly using zero as a placeholder for values that are undefined
+pub(crate) fn finish_interpretation<
+    'a,
+    I: Iterator<Item = &'a ir::Assignment>,
+    P: Into<RcOrConst<ir::Port>>,
+>(
+    mut env: InterpreterState,
+    done_signal: Option<P>,
+    assigns: I,
+) -> InterpreterResult<InterpreterState> {
+    let done_signal: Option<RcOrConst<ir::Port>> =
+        done_signal.map(|x| x.into());
+    // replace port values for all the assignments
+    let assigns = assigns.collect::<Vec<_>>();
+
+    for &ir::Assignment { dst, .. } in &assigns {
+        env.insert(
+            &dst.borrow() as &ir::Port as ConstPort,
+            Value::zeroes(dst.borrow().width as usize),
+        );
+    }
+
+    let cells = get_dest_cells(
+        assigns.iter().copied(),
+        done_signal.as_ref().map(|x| x.get_rrc()).flatten(),
+    );
+
+    if let Some(done_signal) = done_signal {
+        env.insert(done_signal.as_raw(), Value::bit_low());
+    }
+
+    eval_prims(&mut env, cells.iter(), true)?;
+
+    Ok(env)
 }
