--- interp/src/interpreter/steppers/mod.rs
@@ -1,7 +1,0 @@
-mod component_interpreter;
-mod control_interpreter;
-mod group_interpreter;
-
-pub use component_interpreter::ComponentInterpreter;
-pub use control_interpreter::{Interpreter, InvokeInterpreter};
-pub use group_interpreter::AssignmentInterpreter;
